+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This document provides a log of changes. It is currently maintained by hand.
 - [prefab][mongoose-module] Fixed `getModel` returning the schema instead of the model (yikes!)
 - [prefab][mongoose-module] Added setup guards to `registerModel` and `getModel`
 - [prefab][express-module] Fixed http request type proxy methods not accepting a path (yikes!)
+- [prefab][cluster-module] Introduced a prototype prefab for multithreading with typings
 
 ## v0.3