+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This document provides a log of changes. It is currently maintained by hand.
 ## (master)
 
 - Added methods to extract dependencies from SystemState and Module (for debugging purposes for example)
+- [prefab][express-module] Fixed middleware option TS definitions requiring both `before` and `after` properties
 
 ## v0.4